requirements.txt
customtkinter>=5.2.0
pydantic>=2.0.0
pyyaml>=6.0
